+++ lib/mutant/mutator/node/send.rb
@@ -163,6 +163,48 @@ module Mutant
           
           handle(Rubinius::AST::SendWithArguments)
 
+          class BinaryOperatorMethod < Node
+
+          private
+
+            # Emit mutations
+            #
+            # @return [undefined]
+            #
+            # @api private
+            #
+            def dispatch
+              emit_left_mutations
+              emit_right_mutations
+            end
+
+            # Emit left mutations
+            #
+            # @return [undefined]
+            #
+            # @api private
+            #
+            def emit_left_mutations
+              emit_attribute_mutations(:receiver)
+            end
+
+            # Emit right mutations
+            #
+            # @return [undefined]
+            #
+            # @api private
+            #
+            def emit_right_mutations
+              right = node.arguments.array.first
+              Mutator.each(right).each do |mutated|
+                dup = dup_node
+                dup.arguments.array[0] = mutated
+                emit(dup)
+              end
+            end
+
+          end
+
         private
 
           # Emit mutations
@@ -174,6 +216,37 @@ module Mutant
           def dispatch
             super
             emit_call_remove_mutation
+            emit_argument_mutations
+          end
+
+          # Test if message is a binary operator
+          #
+          # @return [true]
+          #   if message is a binary operator
+          #
+          # @return [false]
+          #   otherwise
+          #
+          # @api private
+          #
+          def binary_operator?
+            Mutant::BINARY_METHOD_OPERATORS.include?(node.name)
+          end
+
+          # Emit argument mutations
+          #
+          # @api private
+          #
+          # @return [undefined]
+          #
+          # @api private
+          #
+          def emit_argument_mutations
+            if binary_operator?
+              run(BinaryOperatorMethod)
+              return
+            end
+
             emit_attribute_mutations(:arguments)
           end